lambda/auth_manager/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via starlette
boto3==1.42.59
    # via notes-mcp-auth-manager (pyproject.toml)
botocore==1.42.59
    # via
    #   boto3
    #   s3transfer
cffi==2.0.0
    # via cryptography
click==8.3.1
    # via uvicorn
cryptography==46.0.5
    # via
    #   notes-mcp-auth-manager (pyproject.toml)
    #   python-jose
ecdsa==0.19.1
    # via python-jose
fastapi==0.134.0
    # via notes-mcp-auth-manager (pyproject.toml)
h11==0.16.0
    # via uvicorn
idna==3.11
    # via anyio
jmespath==1.1.0
    # via
    #   boto3
    #   botocore
mangum==0.21.0
    # via notes-mcp-auth-manager (pyproject.toml)
pyasn1==0.6.2
    # via
    #   python-jose
    #   rsa
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   notes-mcp-auth-manager (pyproject.toml)
    #   fastapi
pydantic-core==2.41.5
    # via pydantic
pyjwt==2.11.0
    # via notes-mcp-auth-manager (pyproject.toml)
python-dateutil==2.9.0.post0
    # via botocore
python-jose==3.5.0
    # via notes-mcp-auth-manager (pyproject.toml)
rsa==4.9.1
    # via python-jose
s3transfer==0.16.0
    # via boto3
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
starlette==0.52.1
    # via fastapi
typing-extensions==4.15.0
    # via
    #   fastapi
    #   mangum
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
urllib3==2.6.3
    # via botocore
uvicorn==0.41.0
    # via notes-mcp-auth-manager (pyproject.toml)
